Update to latest gumbo-parser
from https://codeberg.org/gumbo-parser/gumbo-parser

+++ src/gumbo/include/gumbo/utf8.h
@@ -62,7 +62,7 @@ typedef struct GumboInternalUtf8Iterator {
   int _current;
 
   // The width in bytes of the current code point.
-  ptrdiff_t _width;
+  int _width;
 
   // The SourcePosition for the current location.
   GumboSourcePosition _pos;
